refactor: セッションIDを元にユーザーIDを取得する処理の関数化
