+++ src/app/mypage/page.tsx
@@ -1,9 +1,16 @@
 import BreadList from "@/components/frame/breadList/BreadList";
 import PageTitle from "@/components/frame/pageTitle/PageTitle";
 import MypageContents from "@/components/mypage/MypageContents/MypageContents";
+import UnauthorizedAccess from "@/components/user/UnauthorizedAccess/UnauthorizedAccess";
 import { AuthProvider } from "@/contexts/Authcontexts";
+import { checkAuth } from "@/utils/apiFunc";
 
 const MyPage = async () => {
+  const userId = await checkAuth();
+  if(!userId) {
+    return <UnauthorizedAccess />
+  }
+
   return (
     <>
       <BreadList
